feat(routes): return reply from redirected url
